extractor-app/requirements.txt
flask
Flask-Session
Pillow
gunicorn

google-api-python-client
google-cloud
google-cloud-vision
google-cloud-translate
